Rename model names of ax as upper case (AX)
