+++ src/transformers/models/AXK1/configuration_AXK1.py
@@ -44,7 +44,7 @@ class AXK1Config(PreTrainedConfig):
     >>> configuration = model.config
     ```"""
 
-    model_type = "axk1"
+    model_type = "AXK1"
     keys_to_ignore_at_inference = ["past_key_values"]
 
     base_model_tp_plan = {
