+++ src/transformers/models/__init__.py
@@ -30,7 +30,8 @@ if TYPE_CHECKING:
     from .audioflamingo3 import *
     from .auto import *
     from .autoformer import *
-    from .ax_k1 import *
+    from .AXK1 import *
+    from .AXK2 import *
     from .aya_vision import *
     from .bamba import *
     from .bark import *
