+++ src/transformers/models/auto/modeling_auto.py
@@ -55,7 +55,8 @@ MODEL_MAPPING_NAMES = OrderedDict(
         ("audioflamingo3", "AudioFlamingo3ForConditionalGeneration"),
         ("audioflamingo3_encoder", "AudioFlamingo3Encoder"),
         ("autoformer", "AutoformerModel"),
-        ("axk1", "AXK1Model"),
+        ("AXK1", "AXK1Model"),
+        ("AXK2", "AXK2Model"),
         ("aya_vision", "AyaVisionModel"),
         ("bamba", "BambaModel"),
         ("bark", "BarkModel"),
@@ -610,7 +611,8 @@ MODEL_FOR_CAUSAL_LM_MAPPING_NAMES = OrderedDict(
         ("apertus", "ApertusForCausalLM"),
         ("arcee", "ArceeForCausalLM"),
         ("aria_text", "AriaTextForCausalLM"),
-        ("axk1", "AXK1ForCausalLM"),
+        ("AXK1", "AXK1ForCausalLM"),
+        ("AXK2", "AXK2ForCausalLM"),
         ("bamba", "BambaForCausalLM"),
         ("bart", "BartForCausalLM"),
         ("bert", "BertLMHeadModel"),
@@ -1228,7 +1230,8 @@ MODEL_FOR_SEQUENCE_CLASSIFICATION_MAPPING_NAMES = OrderedDict(
         # Model for Sequence Classification mapping
         ("albert", "AlbertForSequenceClassification"),
         ("arcee", "ArceeForSequenceClassification"),
-        ("axk1", "AXK1ForSequenceClassification"),
+        ("AXK1", "AXK1ForSequenceClassification"),
+        ("AXK2", "AXK2ForSequenceClassification"),
         ("bart", "BartForSequenceClassification"),
         ("bert", "BertForSequenceClassification"),
         ("big_bird", "BigBirdForSequenceClassification"),
@@ -1459,7 +1462,8 @@ MODEL_FOR_TOKEN_CLASSIFICATION_MAPPING_NAMES = OrderedDict(
         ("albert", "AlbertForTokenClassification"),
         ("apertus", "ApertusForTokenClassification"),
         ("arcee", "ArceeForTokenClassification"),
-        ("axk1", "AXK1ForTokenClassification"),
+        ("AXK1", "AXK1ForTokenClassification"),
+        ("AXK2", "AXK2ForTokenClassification"),
         ("bert", "BertForTokenClassification"),
         ("big_bird", "BigBirdForTokenClassification"),
         ("biogpt", "BioGptForTokenClassification"),
